requirements.txt
# Backend requirements for GearUp Repairs (Railway-compatible)

fastapi==0.110.1
uvicorn[standard]==0.25.0
pymongo==4.5.0
python-dotenv==1.1.1
